backend/pyproject.toml
[project]
name = "confidential-translator"
version = "0.1.0"
description = "A fully offline, privacy-first document translation system powered by local LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "pymupdf>=1.25.0",
    "pytesseract>=0.3.13",
    "pillow>=11.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.0",
    "langdetect>=1.0.9",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "psutil>=7.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "respx>=0.22.0",
    "pyinstaller>=6.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
